+++ Source/Stuff/cuda_image_creator.cu
@@ -164,8 +164,8 @@ CUDAImageCreator::compile_differences(std::vector<int> &truncated_lin_width, std
         if (x < 0 || x >= dc->simulation_width) { continue;}
         for (auto y: truncated_lin_height) {
             if (y < 0 || y >= dc->simulation_height) { continue;}
-            host_differences.emplace_back(Differences{static_cast<uint16_t>(x),
-                                                      static_cast<uint16_t>(y),
+            host_differences.emplace_back(Differences{static_cast<uint32_t>(x),
+                                                      static_cast<uint32_t>(y),
                                                       dc->second_simulation_grid[x + y * dc->simulation_width].type,
                                                       dc->second_simulation_grid[x + y * dc->simulation_width].rotation,
             });
